# docs/swagger.yaml
basePath: /v1
definitions:
...
  enum.DestinationStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Editing
    - Confirm
    - Deleted
  model.Compound:
    properties:
      commune:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
  model.Destination:
    properties:
      arriveTime:
        type: string
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      lat:
        type: number
      lng:
        type: number
      location:
        $ref: '#/definitions/model.Location'
      locationId:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        description: đơn vị là phút
        type: integer
      updatedTime:
        type: string
    type: object
...
  model.Location:
    properties:
      address:
        type: string
      compound:
        $ref: '#/definitions/model.Compound'
      createdTime:
        type: string
      formattedAddress:
        type: string
      id:
        format: uuid
        type: string
      isVerified:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      placeId:
        type: string
      placeType:
        $ref: '#/definitions/model.PlaceType'
      placeTypeId:
        type: string
      updatedTime:
        type: string
    type: object
  model.PlaceType:
    properties:
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      typeName:
        type: string
      updatedTime:
        type: string
    type: object
...
  request.AddDraftDestinationRequest:
    properties:
      destinationId:
        format: uuid
        type: string
    required:
    - destinationId
    type: object
...
  request.ForkLocationRequest:
    properties:
      arriveTime:
        type: string
      name:
        type: string
      stayOverTime:
        type: integer
    type: object
...
  request.UpdateDestinationRequest:
    properties:
      arriveTime:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        type: integer
    type: object
  request.UpdateGroupRequest:
    properties:
      description:
        maxLength: 4000
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      policy:
        maxLength: 4000
        type: string
    type: object
...
  response.CreateDestinationResponse:
    properties:
      destinationId:
        type: string
      locationId:
        type: string
      placeId:
        type: string
    type: object
  response.DraftDestinationsResponse:
    properties:
      destinations:
        items:
          $ref: '#/definitions/response.StopResponse'
        type: array
    type: object
...
  response.GroupResponse:
    properties:
      createdTime:
        type: string
      description:
        type: string
      id:
        format: uuid
        type: string
      myRole:
        example: owner
        type: string
      name:
        type: string
      ownerId:
        format: uuid
        type: string
      policy:
        type: string
      updatedTime:
        type: string
    type: object
...
  response.StopResponse:
    properties:
      destinationId:
        format: uuid
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      orderInBranch:
        type: integer
    type: object
...
  response.TripInviteLinkResponse:
    properties:
      joinPath:
        example: /v1/trips/join/550e8400-e29b-41d4-a716-446655440000
        type: string
      token:
        type: string
    type: object
...
  share.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        example: invalid input
        type: string
      status:
        example: 400
        type: integer
    type: object
info:
  contact: {}
  description: 'Backend API for Road2D (team Vandra): collaborative multi-branch ride
    planning and group-ride safety in Vietnam. JWT auth, Goong Maps playground, trip
    graph models.'
  title: Road2D
  version: "1.0"
paths:
...
  /groups/{groupId}:
    delete:
      description: Owner only. Members cascade-delete with the group. Trips are not
        tied to a group.
      parameters:
      - description: Group UUID
        format: uuid
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Group deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - groups
    get:
      description: Returns name and description. If the caller is an active member,
        myRole is set (owner/admin/member). Non-members still receive public name/description
        so they can request to join; roster is never included here.
      parameters:
      - description: Group UUID
        format: uuid
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group
      tags:
      - groups
    patch:
      consumes:
      - application/json
      description: Owner or admin may change name, description, and/or policy (group
        rules text). Members cannot update group info. Send only the fields to change.
      parameters:
      - description: Group UUID
        format: uuid
        in: path
        name: groupId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - groups
...
  /planning/destination/{destinationId}:
    delete:
      description: Deletes a destination that is still editing. A pin that sits only
        on draft branches is removed immediately. A pin on a route branch is stripped
        first; the delete is rejected when that would disconnect the graph or when
        a trip that is no longer planning still uses the pin on a route. Travels that
        pointed at the pin are removed with the row.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Destination deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete destination
      tags:
      - planning
    get:
      description: Returns a destination. When locationId is set, lat/lng are locked
        to that location.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get destination
      tags:
      - planning
    put:
      consumes:
      - application/json
      description: Updates name, arrive time, stay, and status while the destination
        is editing. lat/lng are applied only when locationId is empty; a forked or
        verified pin keeps the linked location's coordinates.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      - description: Fields to store
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDestinationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update destination
      tags:
      - planning
  /planning/fork/{locationId}:
    post:
      consumes:
      - application/json
      description: Creates an editing destination from a verified location. lat/lng
        and the location link are copied from the location. Optional name overrides
        the location name.
      parameters:
      - description: Location UUID
        format: uuid
        in: path
        name: locationId
        required: true
        type: string
      - description: Optional name, arrive time, and stay
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ForkLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CreateDestinationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fork location
      tags:
      - planning
...
  /trips/{tripId}/draft:
    get:
      description: Returns every destination stored on the trip's hidden draft branch,
        in inbox order. This branch is not routed and is omitted from GET /trips/{tripId}/graph
        and GET /trips/{tripId}/travels. Any active member may call this.
      parameters:
      - description: Trip id
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DraftDestinationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List draft destinations
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Appends one destination to the trip's hidden draft branch. The
        pin must already exist, usually from POST /planning/fork/{locationId}. Only
        while the trip is planning. This does not change the route graph or the stored
        travels.
      parameters:
      - description: Trip id
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Destination to park
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AddDraftDestinationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DraftDestinationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a draft destination
      tags:
      - trips
...
  /trips/{tripId}/invite-link:
    post:
      description: Leader or admin. Returns the current invite token and join path.
        Set rotate=true to invalidate the previous link.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Mint a new token and drop the old one
        in: query
        name: rotate
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripInviteLinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create trip invite link
      tags:
      - trips
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  PlaygroundKey:
    description: Shared secret from GOONG_PLAYGROUND_KEY. Scalar → Authorize.
    in: header
    name: X-Playground-Key
    type: apiKey
swagger: "2.0"